LC-combination-sum-iv [Runtime: N/A; Memory: N/A]

+++ array/combination-sum-iv.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int combinationSum4(vector<int>& nums, int target) {
-        vector<long long>dp(target+1, 0);
+        vector<unsigned int>dp(target+1, 0);
         dp[0] = 1;
 
         for(int sum = 1; sum<=target; sum++){
